pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "dependency-injector>=4.42.0",
    "fastapi[standard]>=0.115.3",
    "loguru>=0.7.2",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.9.0",
    "sqlalchemy>=2.0.36",
    "pycaret>=3.3.2",
    "psycopg2-binary>=2.9.10",
    "sqlmodel>=0.0.22",
    "bcrypt>=4.2.1",
    "psycopg2>=2.9.10",
    "google-cloud-aiplatform>=1.74.0",
    "langchain>=0.3.10",
    "langchain-google-vertexai>=2.0.8",
    "langchain-community>=0.3.10",
    "google-cloud-storage>=2.19.0",
]
